- replacing the numeric(38) and numeric(1000) stuff with bigint (cherry pick merge from trunk 647+648)

+++ Annis-Service/src/main/sql/staging_area.sql
@@ -17,12 +17,12 @@ DROP TABLE IF EXISTS _text;
 
 CREATE :tmp TABLE _corpus
 (
-	id 			numeric(38) NOT NULL,		-- primary key
+	id 			bigint NOT NULL,		-- primary key
 	name		varchar(100) NOT NULL,		-- unique name
 	type		varchar(100) NOT NULL,		-- CORPUS, DOCUMENT, SUBDOCUMENT (not used)
 	version		varchar(100),				-- version number (not used)
-	pre			numeric(38)	NOT NULL,		-- pre and post order of the corpus tree
-	post		numeric(38)	NOT NULL
+	pre			bigint	NOT NULL,		-- pre and post order of the corpus tree
+	post		bigint	NOT NULL
 );
 
 -- corpus annotations
@@ -30,7 +30,7 @@ CREATE :tmp TABLE _corpus
 
 CREATE :tmp TABLE _corpus_annotation
 (
-	corpus_ref	numeric(38) NOT NULL,		-- foreign key to _corpus.id
+	corpus_ref	bigint NOT NULL,		-- foreign key to _corpus.id
 	namespace	varchar(100),
 	name		varchar(1000) NOT NULL,
 	value		varchar(2000)
@@ -40,7 +40,7 @@ CREATE :tmp TABLE _corpus_annotation
 -- stores each source text in its entirety
 CREATE :tmp TABLE _text
 (
-	id 		numeric(38) NOT NULL,			-- primary key
+	id 		bigint NOT NULL,			-- primary key
 	name	varchar(1000),					-- name (not used)
 	text 	text							-- text contents (not used)
 );
@@ -52,9 +52,9 @@ CREATE :tmp TABLE _text
 -- can be tokens (token_index NOT NULL, span NOT NULL)
 CREATE :tmp TABLE _node
 (
-	id 				numeric(38)	NOT NULL,	-- primary key
-	text_ref 		numeric(38) NOT NULL,	-- foreign key to _text.id
-	corpus_ref		numeric(38) NOT NULL,	-- foreign key to _corpus.id
+	id 				bigint	NOT NULL,	-- primary key
+	text_ref 		bigint NOT NULL,	-- foreign key to _text.id
+	corpus_ref		bigint NOT NULL,	-- foreign key to _corpus.id
 	namespace		varchar(100),			-- namespace (not used)
 	name 			varchar(100) NOT NULL,	-- name (not used)
 	"left" 			integer NOT NULL,		-- start of covered substring in _text.text (inclusive)
@@ -69,7 +69,7 @@ CREATE :tmp TABLE _node
 -- have a name
 CREATE :tmp TABLE _component
 (
-	id			numeric(38) NOT NULL,		-- primary key
+	id			bigint NOT NULL,		-- primary key
 	type		char(1),					-- edge type: c, d, P, NULL
 	namespace	varchar(255),				-- namespace (not used)
 	name		varchar(255)
@@ -80,18 +80,18 @@ CREATE :tmp TABLE _component
 -- component and rank together model edges in the annotation graph
 CREATE :tmp TABLE _rank
 (
-	pre				numeric(38)	NOT NULL,	-- pre and post order of the annotation ODAG
-	post			numeric(38)	NOT NULL,
-	node_ref		numeric(38)	NOT NULL,	-- foreign key to _node.id
-	component_ref	numeric(38) NOT NULL,	-- foreign key to _component.id
-	parent			numeric(38) NULL		-- foreign key to _rank.pre, NULL for root nodes
+	pre				bigint	NOT NULL,	-- pre and post order of the annotation ODAG
+	post			bigint	NOT NULL,
+	node_ref		bigint	NOT NULL,	-- foreign key to _node.id
+	component_ref	bigint NOT NULL,	-- foreign key to _component.id
+	parent			bigint NULL		-- foreign key to _rank.pre, NULL for root nodes
 );
 
 -- node annotations
 -- unique combinantion of node_ref, namespace and name
 CREATE :tmp TABLE _node_annotation
 (
-	node_ref	numeric(38)	NOT NULL,		-- foreign key to _node.id
+	node_ref	bigint	NOT NULL,		-- foreign key to _node.id
 	namespace	varchar(150),
 	name		varchar(150) NOT NULL,
 	value		varchar(1500)
@@ -102,7 +102,7 @@ CREATE :tmp TABLE _node_annotation
 -- unique combinantion of node_ref, namespace and name
 CREATE :tmp TABLE _edge_annotation
 (
-	rank_ref		numeric(38)	NOT NULL,	-- foreign key to _rank.pre
+	rank_ref		bigint	NOT NULL,	-- foreign key to _rank.pre
 	namespace		varchar(1000),
 	name			varchar(1000) NOT NULL,
 	value			varchar(1000)
@@ -118,6 +118,6 @@ CREATE :tmp TABLE _resolver_vis_map
   "element"    varchar(4), -- the type of the entry: node | edge
   "vis_type"   varchar(100) NOT NULL, -- the abstract type of visualization: tree, discourse, grid, ...
   "display_name"   varchar(100) NOT NULL, -- the name of the layer which shall be shown for display
-  "order" numeric default '0', -- the order of the layers, in which they shall be shown
+  "order" bigint default '0', -- the order of the layers, in which they shall be shown
   "mappings" varchar(100)			    
 );
